refactor(backend): share the current-user accessor across hosts
Move ICurrentUserAccessor and the claims->CurrentUser parsing into
Kalandra.Infrastructure (a pure CurrentUserFactory, no ASP.NET dependency) so the
coming MCP host reads identity from a Supabase token exactly as the API does. Each
host keeps a thin IHttpContextAccessor wrapper that delegates to the factory.
