+++ backend/src/Kalandra.Api/Features/Mcp/BlogMcpTools.cs
@@ -1,6 +1,5 @@
 using System.ComponentModel;
 using Kalandra.Api.Features.Mcp.Contracts;
-using Kalandra.Api.Infrastructure.Auth;
 using Kalandra.Blog;
 using Kalandra.Blog.Commands;
 using Kalandra.Blog.Contracts;
@@ -8,6 +7,7 @@ using Kalandra.Blog.Entities;
 using Kalandra.Blog.Events;
 using Kalandra.Blog.Feed;
 using Kalandra.Blog.Queries;
+using Kalandra.Infrastructure.Auth;
 using Kalandra.JobOffers.Queries;
 using ModelContextProtocol;
 using ModelContextProtocol.Server;
