+++ backend/src/Kalandra.Api/Features/Mcp/JobOfferMcpTools.cs
@@ -1,5 +1,5 @@
 using System.ComponentModel;
-using Kalandra.Api.Infrastructure.Auth;
+using Kalandra.Infrastructure.Auth;
 using Kalandra.JobOffers.Contracts;
 using Kalandra.JobOffers.Commands;
 using Kalandra.JobOffers.Entities;
